+++ imap_tools/__init__.py
@@ -1,7 +1,7 @@
 from .query import Q, AND, OR, NOT, Header, A, O, N, H
-from .mailbox import *
-from .message import *
-from .folder import *
+from .mailbox import BaseMailBox, MailBox, MailBoxUnencrypted
+from .message import MailMessage, Attachment, MailMessageFlags
+from .folder import MailBoxFolderManager, MailBoxFolderStatusOptions
 from .errors import *
 
-__version__ = '0.21.0'
+__version__ = '0.22.0'